Changes to output import/export power
Need to generalize the report outputs across the problem types.  If a
problem has network data it should automatically report network results.
Economic results should be reported when an economic dispatch has
occurred (most cases) .

Going to consider if report data should be sent to excel or database
should be an option.